+++ tutorials/helpers.php
@@ -452,3 +452,5 @@ function colorize(string $text, string $color): string
 }
 
 
+
+